requirements.txt
#telebot==0.0.4
python-telegram-bot==13.13
pyTelegramBotAPI==4.7.0
pandas==1.4.3
pandas-gbq==0.17.8
openai==0.22.1
gunicorn==20.1.0
Flask==2.1.2
Unidecode==1.3.4
google-cloud-storage==2.5.0
google-auth-oauthlib==0.5.2
google-auth==2.10.0
google-cloud==0.34.0
google-cloud-bigquery==3.3.1
jsonschema==4.14.0
prettytable==3.4.1
requests==2.28.1
requests-oauthlib==1.3.1
pyzmq==23.2.1